Project entitled: Effects of pH, Metal Concentration, and Salinity on Percentage of Removal in the Chromium (VI) Metal Ion Biosorption Process by Using Chlorella sp. Microalgae.
Chromium (VI) Metal Ions is one of the heavy metals which are highly toxic. This toxic ions can lead to chronic, acute poisoning, and can cause cancer. The Chlorella sp. microalgae has been widely known to have the ability to survive in harsh environment even in high concentration of heavy metal ions, such as copper (Cu) and using these ions as part of its metabolism. The purpose of these project was to determine and analyze the abilities of Chlorella sp. microalgae in conducting of biosorption process of chromium (VI) metal ions in different conditions.